более проработал функцию по работе и полем адреса
