+++ js/util.js
@@ -30,4 +30,6 @@ const randomArrElements = (source) => {
   return [ ...new Set(arr) ];
 };
 
+
+
 export {getRandomNumber, getRandomFloatingPointNumber, getСalculatingRandomNumber, randomArrElements};
